Move mdns from src/lib/protocols to src/lib
