+++ examples/wifi-echo/server/esp32/main/DeviceCallbacks.cpp
@@ -30,7 +30,7 @@
 #include "WiFiWidget.h"
 #include "esp_heap_caps.h"
 #include "esp_log.h"
-#include <lib/protocols/mdns/Publisher.h>
+#include <lib/mdns/Publisher.h>
 #include <support/CodeUtils.h>
 
 extern "C" {
