+++ src/platform/BUILD.gn
@@ -353,7 +353,7 @@ if (chip_device_platform != "none" && chip_device_platform != "external") {
           "Linux/MdnsImpl.h",
         ]
 
-        public_deps += [ "${chip_root}/src/lib/protocols/mdns:platform_header" ]
+        public_deps += [ "${chip_root}/src/lib/mdns:platform_header" ]
 
         public_configs += [ ":avahi_client_config" ]
       }
